Reorganize WFM model (Phase1)
