+++ client/wfm/src/index.ts
@@ -1,10 +1,9 @@
-export { WorkOrder } from './workorder/WorkOrder';
-export { WorkFlow } from './workflow/WorkFlow';
-export { WorkOrderResult } from './result/WorkOrderResult';
-export { StepResult } from './result/StepResult';
-export { Step } from './step/Step';
+export { WorkOrder } from './model/WorkOrder';
+export { WorkFlow } from './model/WorkFlow';
+export { WorkOrderResult } from './model/WorkOrderResult';
+export { StepResult } from './model/StepResult';
+export { Step } from './model/Step';
 export { STATUS } from './status';
 export { WfmService } from './service/WfmService';
-export { ResultService } from './service/ResultService';
 export { UserService } from './service/UserService';
 export { DataService } from './service/DataService';
