--- client/wfm/src/service/ResultService.ts
@@ -1,7 +1,0 @@
-import * as Promise from 'bluebird';
-import { WorkOrderResult } from '../result/WorkOrderResult';
-import { DataService } from './DataService';
-
-export interface ResultService extends DataService<WorkOrderResult> {
-  readByWorkorder(workorderId: string): Promise<WorkOrderResult | undefined>;
-}
